jwt_auth/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
PyJWT==2.10.1
APScheduler==3.11.0
Werkzeug==3.1.3
